+++ src/renderer/components/Caption.jsx
