+++ src/renderer/components/DeleteModal.jsx
