+++ src/renderer/components/Footer.jsx
