+++ src/renderer/components/HostSelector.jsx
