+++ src/renderer/components/MenuIcon.jsx
