+++ src/renderer/components/Navigation/SearchBar.jsx
