+++ src/renderer/components/Paginator.jsx
