+++ src/renderer/components/PaginatorButton/PrevPageButton.jsx
