+++ src/renderer/components/Player/ImageSelector/DeleteButton.jsx
