+++ src/renderer/components/ProjectCard.jsx
