+++ src/renderer/components/ProjectDetail/ProjectCard.jsx
