+++ src/renderer/components/ProjectEditForm/AnnotationTool/AnnotationImage.jsx
