+++ src/renderer/components/ProjectEditForm/AnnotationTool/Rectangle.jsx
