+++ src/renderer/components/ProjectEditForm/CaptionsField.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import CaptionField from './CaptionField';
+import CaptionField from './CaptionField.jsx';
 
 import {
   EditBanner,
@@ -19,17 +19,20 @@ const CaptionsField = ({ figures, contentType, handleCaptionsChange, onAddCaptio
     <StyledCaptionsField>
       <EditBanner className="edit">Captions</EditBanner>
       <StyledCaptionFieldHeader>
-        <CaptionFieldSpan text="start" contentType={contentType}>start(h:m:s)</CaptionFieldSpan>
-        <CaptionFieldSpan text="end" contentType={contentType}>end(h:m:s)</CaptionFieldSpan>
-        <CaptionFieldSpan text="text" contentType={contentType}>caption</CaptionFieldSpan>
+        <CaptionFieldSpan text="start" contentType={contentType}>
+          start(h:m:s)
+        </CaptionFieldSpan>
+        <CaptionFieldSpan text="end" contentType={contentType}>
+          end(h:m:s)
+        </CaptionFieldSpan>
+        <CaptionFieldSpan text="text" contentType={contentType}>
+          caption
+        </CaptionFieldSpan>
         <CaptionFieldSpan>delete</CaptionFieldSpan>
       </StyledCaptionFieldHeader>
       {figures.map((figure, figureIndex) => {
         return (
-          <CaptionFieldWrapper
-            key={`figure_${figure.id}_captions`}
-            figureWillBeDeleted={figure._destroy}
-          >
+          <CaptionFieldWrapper key={`figure_${figure.id}_captions`} figureWillBeDeleted={figure._destroy}>
             <CaptionFieldIndex> {`${contentType}#${figureIndex + 1}`}</CaptionFieldIndex>
             <CaptionFieldList>
               {figure.captions.map((caption, index) => {
@@ -45,12 +48,8 @@ const CaptionsField = ({ figures, contentType, handleCaptionsChange, onAddCaptio
                 );
               })}
             </CaptionFieldList>
-            <AddCaptionButton
-              className="addCaptionButton"
-              onClick={onAddCaptionButtonClick}
-              data-index={figureIndex}
-            >
-                            add caption
+            <AddCaptionButton className="addCaptionButton" onClick={onAddCaptionButtonClick} data-index={figureIndex}>
+              add caption
             </AddCaptionButton>
           </CaptionFieldWrapper>
         );
