+++ src/renderer/components/ProjectManager.jsx
@@ -2,14 +2,14 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import Debug from 'debug';
 
-import LeftNav from './Navigation/LeftNav';
-import RightNav from './Navigation/Rightnav';
-import CenterNav from './Navigation/CenterNav';
-import Footer from './Footer';
-import BackButton from './BackButton';
-import NextPageButton from './PaginatorButton/NextPageButton';
-import PrevPageButton from './PaginatorButton/PrevPageButton';
-import ProjectSettings from './ProjectSettings';
+import LeftNav from './Navigation/LeftNav.jsx';
+import RightNav from './Navigation/Rightnav.jsx';
+import CenterNav from './Navigation/CenterNav.jsx';
+import Footer from './Footer.jsx';
+import BackButton from './BackButton.jsx';
+import NextPageButton from './PaginatorButton/NextPageButton.jsx';
+import PrevPageButton from './PaginatorButton/PrevPageButton.jsx';
+import ProjectSettings from './ProjectSettings.jsx';
 
 import {
   PageFrame,
