+++ src/renderer/components/ProjectSettings.jsx
