+++ src/renderer/components/Visualizer/ProcessVisualizer.jsx
