+++ src/renderer/components/Visualizer/TagPerStep.jsx
